Adapt nextcloud share manager to new method signature
Also remove the test for UpdateReceivedShare which can not be tested
anymore with the new signature. The ReceivedShare never held the display
name that's being tested so the test only passed on the data from the
update field, but since the method only takes the actual received share
now this is no longer possible.